pyproject.toml
[tool.poetry]
name = "doughman"
version = "0.1.0"
description = ""
authors = ["Robert J. Berger <rberger@ibd.com>"]
readme = "README.md"
packages = [{include = "doughman", from = "src"}]

[tool.poetry.scripts]
#doughman = "doughman:cli.cli"
doughman = "doughman:tui.tui_run"

[tool.poetry.dependencies]
python = "^3.11"
# vl53_400_lib = "^0.1"
vl53_400_Lib = { path = "../lidar/vl53_400_lib/", develop = true }
click = "^8.1.7"
textual = "^0.41"
pyserial = "^3.5"
pyserial-asyncio = "^0.6"

[tool.poetry.group.dev.dependencies]
textual-dev = "*"
ruff = "^0.0.287"
pytest = "^7.4.1"
black = "^23.7.0"
coverage = "^7.3.0"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120

[tool.mypy]
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = false
disallow_subclassing_any = true
# platform
python_version = "3.11"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
